parser_service/requirements-version.txt
beautifulsoup4==4.12.3
ftfy==6.3.1
grpcio==1.69.0
grpcio_reflection==1.69.0
huggingface_hub==0.27.1
marker==2.1.5
numpy==2.2.1
Pillow==11.1.0
protobuf==5.29.3
redis==5.2.1
Requests==2.32.3
torch==2.5.1
tqdm==4.67.1
